[7X] Skip ssh if segments are on the same host with coordinator. (#16812)
We can skip ssh connections if segments are on the same host with the
coordinator. We can save some time when executing commands. E.g., with
this patch, we don't ssh connection when creating the demo cluster and
the creation time can be reduced from 18s to 10s on my laptop. Besides,
we can skip setting up ssh keys for deploying the demo cluster.